refactor(260718-fxm): reorganize events package by trading domain
- Add portfolio.py (PortfolioUpdateEvent), screener.py (ScreenerEvent),
  strategy.py (StrategyCommandEvent + 9 factories), feed.py (BarsLoaded/BarsLoadFailed)
- Trim market.py to TimeEvent + BarEvent; move UniverseUpdateEvent into universe.py
- Merge OrderAckEvent into order.py; delete ack.py
- Update barrel to source each name from its new module (public names unchanged)
- Repoint 6 direct-submodule importers of UniverseUpdateEvent market -> universe
