+++ itrader/events_handler/events/__init__.py
@@ -14,30 +14,40 @@ from .base import Event
 from .market import (
     TimeEvent,
     BarEvent,
-    PortfolioUpdateEvent,
-    ScreenerEvent,
-    UniverseUpdateEvent,
 )
 
-# Universe / live control-plane events (D-03/D-04/D-06/D-09)
+# Portfolio snapshot events (D-07)
+from .portfolio import PortfolioUpdateEvent
+
+# Screener events (D-screener)
+from .screener import ScreenerEvent
+
+# Universe events (D-04/D-06)
 from .universe import (
-    BarsLoaded,
-    BarsLoadFailed,
+    UniverseUpdateEvent,
     UniversePollEvent,
-    StrategyCommandEvent,
 )
 
 # Signal events
 from .signal import SignalEvent
 
-# Order events
-from .order import OrderEvent
+# Strategy control-plane command events (D-08/D-09)
+from .strategy import StrategyCommandEvent
 
-# Order-ack events (D-06)
-from .ack import OrderAckEvent
+# Order events + order-ack events (D-06)
+from .order import (
+    OrderEvent,
+    OrderAckEvent,
+)
 
 # Fill events
 from .fill import FillEvent
+
+# Warmup bulk-transport events (D-03/D-04)
+from .feed import (
+    BarsLoaded,
+    BarsLoadFailed,
+)
 
 # Error events (D-06)
 from .error import (
