+++ itrader/events_handler/events/market.py
@@ -1,9 +1,9 @@
 """
-Market-side events: clock ticks, bars, portfolio snapshots and screener
-results. All frozen ``Event`` subclasses (M3-01).
+Market-side events: clock ticks and OHLCV bars. All frozen ``Event``
+subclasses (M3-01).
 """
 
-from typing import Any, ClassVar
+from typing import ClassVar
 
 from itrader.core.bar import Bar
 from itrader.core.enums import EventType
@@ -45,81 +45,6 @@ class BarEvent(Event, frozen=True, kw_only=True, gc=False):
 
     type: ClassVar[EventType] = EventType.BAR
     bars: dict[str, Bar]
-
-    def __str__(self) -> str:
-        return f"{self.type}, Time: {self.time}"
-
-    def __repr__(self) -> str:
-        return str(self)
-
-
-class PortfolioUpdateEvent(Event, frozen=True, kw_only=True, gc=False):
-    """
-    Periodic snapshot of all portfolios' state, generated by the
-    PortfolioHandler. ``portfolios`` keeps its current documented dict
-    shape (D-07 — no new DTOs this phase).
-    """
-
-    type: ClassVar[EventType] = EventType.UPDATE
-    portfolios: dict[str, Any]
-
-    def __str__(self) -> str:
-        return f"{self.type}, Time: {self.time}"
-
-    def __repr__(self) -> str:
-        return str(self)
-
-
-class ScreenerEvent(Event, frozen=True, kw_only=True, gc=False):
-    """
-    Screener event generated from a Screener object.
-    This is received by the Strategy handler object
-    that update the symbol to trade of the subscribed
-    strategies.
-
-    Current fields kept unchanged (frozen structurally; D-screener owns
-    the semantics).
-    """
-
-    type: ClassVar[EventType] = EventType.SCREENER
-    screener_id: str
-    screener_name: str
-    subscribed_strategies: list[str]
-    tickers: list[str]
-
-    def __str__(self) -> str:
-        return f"{self.type} ({self.screener_name})"
-
-    def __repr__(self) -> str:
-        return str(self)
-
-
-class UniverseUpdateEvent(Event, frozen=True, kw_only=True, gc=False):
-    """Dynamic-universe membership change notification (D-04).
-
-    The "dispose" notification for mid-run universe membership: after a poll
-    resolves adds/removes, ``Universe.apply`` mutates membership and the poll
-    handler emits ONE ``UniverseUpdateEvent`` carrying the resulting delta.
-
-    DISTINCT from ``ScreenerEvent`` (the "propose" seam — screeners propose,
-    membership disposes): this event announces a membership change that already
-    happened, not a proposal. Kept a separate type so the two seams never
-    conflate on the queue.
-
-    Payload contract:
-
-    - ``added`` / ``removed`` are ``tuple[str, ...]`` (immutable payload on the
-      frozen struct) — the symbols that entered / left the universe.
-    - ``time`` / ``event_id`` / ``created_at`` are inherited from ``Event``
-      (business time, UUIDv7 identity — never wall clock).
-
-    This plan ships NO consumers: the ``_routes`` entry is explicit-empty and
-    live consumers are wired live-only in plan 05 (backtest stays inert).
-    """
-
-    type: ClassVar[EventType] = EventType.UNIVERSE_UPDATE
-    added: tuple[str, ...]
-    removed: tuple[str, ...]
 
     def __str__(self) -> str:
         return f"{self.type}, Time: {self.time}"
