+++ itrader/events_handler/events/portfolio.py
@@ -1,0 +1,27 @@
+"""
+Portfolio snapshot event: the PortfolioHandler's periodic all-portfolios
+state broadcast (D-07). A frozen ``Event`` subclass (M3-01).
+"""
+
+from typing import Any, ClassVar
+
+from itrader.core.enums import EventType
+
+from .base import Event
+
+
+class PortfolioUpdateEvent(Event, frozen=True, kw_only=True, gc=False):
+    """
+    Periodic snapshot of all portfolios' state, generated by the
+    PortfolioHandler. ``portfolios`` keeps its current documented dict
+    shape (D-07 — no new DTOs this phase).
+    """
+
+    type: ClassVar[EventType] = EventType.UPDATE
+    portfolios: dict[str, Any]
+
+    def __str__(self) -> str:
+        return f"{self.type}, Time: {self.time}"
+
+    def __repr__(self) -> str:
+        return str(self)
